Video Resolution Fix!
This took a while to figure out, but the solution ended up being pretty simple. There is now a border that follows the camera around, and the stage has been re-sized to fit the camera. Now, no matter the resolution of the computer playing the level, the edge of the screen will ALWAYS cause the camera to switch to the next area!!!

This is a pretty drastic change, so I put it on a separate branch, just in case. It will probably need some minor adjustments to get feeling exactly right, but it's basically set. This branch should also have all of the changes from the Energy System Commit.