Hole Teleport System Fully Implemented and Added as Prefabs
Should be easy enough to make adjustments for play easily enough. I skipped making entering the hole a button prompt, I've got enough work to do here as it is.

I also removed the colliders from the numbers in this commit; they somehow manage to still get in the way of movement sometimes.